# main.go
package main

import (
	"log"
	"os"

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/template/html/v2"
	"github.com/joho/godotenv"

	jwtware "github.com/gofiber/jwt/v4"

	"github.com/gofiber/swagger"

	// docs are generated by Swag CLI, you have to import them.
	// replace with your own docs folder, usually "github.com/username/reponame/docs"
	_ "github.com/alohadancemeow/go-api-demo/docs"
)

// Book struct to hold book data
type Book struct {
	ID     int    `json:"id"`
	Title  string `json:"title"`
	Author string `json:"author"`
}

var books []Book // Slice to store books

// Dummy user for example
type User = struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

var member = User{
	Email:    "user@email.com",
	Password: "123456",
}

// @title Book API /w Fiber
// @version 1.0
// @description This is a sample swagger for Fiber
// @host localhost:3000
// @BasePath /
// @schemes http
// @securityDefinitions.apikey ApiKeyAuth
// @in header
// @name Authorization
func main() {

	// Load environment variables from .env file
	err := godotenv.Load()
	if err != nil {
		log.Fatal("Error loading .env file")
	}

	engine := html.New("./views", ".html")

	// Initialize a new Fiber app
	app := fiber.New(fiber.Config{
		// Pass in Views Template Engine
		Views: engine,
	})

	// Swagger route
	app.Get("/swagger/*", swagger.HandlerDefault) // default

	// Initialize in-memory data
	books = append(books, Book{ID: 1, Title: "1984", Author: "George Orwell"})
	books = append(books, Book{ID: 2, Title: "The Great Gatsby", Author: "F. Scott Fitzgerald"})

	// Public route ["/", "/html-test"]
	app.Get("/", func(c *fiber.Ctx) error {
		// Send a string response to the client
		return c.SendString("Hello, World 👋!")
	})

	// Render HTML template
	app.Get("/html-test", func(c *fiber.Ctx) error {
		// Render index template
		return c.Render("index", fiber.Map{
			"Title":       "Go Fiber Template Example",
			"Description": "An example template",
			"Greeting":    "Hello, World!",
		})
	})

	// Login
	app.Post("/login", login)

	// JWT Middleware
	app.Use(jwtware.New(jwtware.Config{
		SigningKey: jwtware.SigningKey{
			Key: []byte(os.Getenv("JWT_SECRET_KEY")),
		},
	}))

	// Protected route
	app.Use(checkMiddleware)

	// CRUD routes
	app.Get("/books", getBooks)
	app.Get("/book/:id", getBook)
	app.Post("/book", createBook)
	app.Put("/book/:id", updateBook)
	app.Delete("/book/:id", deleteBook)

	// Upload file
	app.Post("/upload", uploadFile)

	// Get environment variables
	app.Get("/env", getEnv)

	// Start the server on port 3000
	log.Fatal(app.Listen(":3000"))
}

func uploadFile(c *fiber.Ctx) error {
	file, err := c.FormFile("image")
	if err != nil {
		return c.Status(fiber.StatusBadRequest).SendString(err.Error())
	}

	err = c.SaveFile(file, "./uploads/"+file.Filename)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).SendString(err.Error())
	}

	return c.SendString("File uploaded successfully!")
}

// get environment variables
func getEnv(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"SECRET_KEY": os.Getenv("SECRET_KEY"),
	})
}